requirements.txt
# Google Cloud ADK and AI Platform
google-adk>=1.0.0
google-cloud-aiplatform>=1.77.0
google-cloud-firestore>=2.16.0
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.16.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0

# Financial Data and Analysis
yfinance>=0.2.28
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0

# Web Framework and Visualization
streamlit>=1.35.0
plotly>=5.20.0
altair>=5.3.0

# Trading APIs
alpaca-py>=0.20.0

# Async and HTTP (FIXED - removed asyncio-compat)
aiohttp>=3.9.0
httpx>=0.27.0

# Data Processing
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# News and Social Media
newsapi-python>=0.2.7
praw>=7.7.1

# Utilities
python-dateutil>=2.9.0
pytz>=2024.1
typing-extensions>=4.11.0

# Development and Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
black>=24.4.0
flake8>=7.0.0
mypy>=1.10.0

# Security
cryptography>=42.0.0
PyJWT>=2.8.0

# Optional: Development tools
jupyter>=1.0.0
notebook>=7.2.0
